Rename the browser-modern utils/index.js to utils/dom-getters.js
