+++ app/ui/browser-modern/util/dom-getters.js
